+++ changelog.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [ ] Write unit and integration tests
 - [ ] Prepare deployment scripts and documentation
 ---
+## [0.2.0] - 2024-14-10
+- added google book api for easier book search
+- made search bars more suer friendly
+
 ## [0.0.2] - 2024-11-10
 - add books form works and updates database
 - log pages read works and goes to log table